refactor(assistant): move conversation/message storage into persistence/ with shims
